+++ src/main/kotlin/nl/info/zac/app/filter/cache/StaticCacheFilter.kt
@@ -3,24 +3,22 @@
  * SPDX-License-Identifier: EUPL-1.2+
  */
 
-package nl.info.zac.util
+package nl.info.zac.app.filter.cache
 
 import jakarta.servlet.Filter
 import jakarta.servlet.FilterChain
-import jakarta.servlet.ServletOutputStream
 import jakarta.servlet.ServletRequest
 import jakarta.servlet.ServletResponse
 import jakarta.servlet.annotation.WebFilter
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import jakarta.servlet.http.HttpServletResponseWrapper
-import java.io.PrintWriter
+import kotlin.time.Duration.Companion.days
 
 /**
  * Overrides Undertow's default `no-cache, no-store` headers on static assets.
  *
  * Undertow writes response headers when [getOutputStream] or [getWriter] is first called — at that
- * point the response is not yet committed, so headers can still be set. A [CacheControlResponseWrapper]
+ * point the response is not yet committed, so headers can still be set. A [ResponseWrapper]
  * intercepts both calls to inject the correct Cache-Control value and suppress the legacy
  * `Pragma` and `Expires` headers Undertow also sets.
  *
@@ -31,10 +29,13 @@ import java.io.PrintWriter
  * - Versioned `/assets/` files with a valid `?v=` MD5 param: `immutable`
  * - `index.html`: `no-cache` — must revalidate so new chunk references are picked up after deploy
  */
-@WebFilter(filterName = "StaticResourceCacheFilter", urlPatterns = ["/*"])
-class StaticResourceCacheFilter : Filter {
+@WebFilter(filterName = "StaticCacheFilter", urlPatterns = ["/*"])
+class StaticCacheFilter : Filter {
 
     companion object {
+        /** Max-age for immutable assets: 1 year in seconds, the conventional maximum for `Cache-Control: immutable`. */
+        private val MAX_AGE_SECONDS = 365.days.inWholeSeconds
+
         /** 8 is Angular's default hash length (hardcoded in `@angular/build`); filename example: main-A1B2C3D4.js */
         private val HASHED_RESOURCE_REGEX = Regex("""-[A-Za-z0-9]{8}\.(js|css)(\.map)?$""")
 
@@ -46,7 +47,7 @@ class StaticResourceCacheFilter : Filter {
         val wrappedResponse = (request as? HttpServletRequest)
             ?.let { httpRequest -> resolveCacheControl(httpRequest) }
             ?.let { cacheControl ->
-                (response as? HttpServletResponse)?.let { CacheControlResponseWrapper(it, cacheControl) }
+                (response as? HttpServletResponse)?.let { ResponseWrapper(it, cacheControl) }
             }
             ?: response
         chain.doFilter(request, wrappedResponse)
@@ -57,53 +58,9 @@ class StaticResourceCacheFilter : Filter {
         return when {
             HASHED_RESOURCE_REGEX.containsMatchIn(path) ||
                 (path.startsWith("/assets/") && MD5_VERSION_REGEX.matches(request.getParameter("v") ?: "")) ->
-                "public, max-age=31536000, immutable"
+                "public, max-age=$MAX_AGE_SECONDS, immutable"
             path == "/" || path == "/index.html" -> "no-cache"
             else -> null
-        }
-    }
-
-    /** Blocks Undertow from setting cache-related headers; injects the correct value on [getOutputStream] and [getWriter]. */
-    private class CacheControlResponseWrapper(
-        response: HttpServletResponse,
-        private val cacheControl: String
-    ) : HttpServletResponseWrapper(response) {
-
-        private fun isCacheHeader(name: String) =
-            name.equals("Cache-Control", ignoreCase = true) ||
-                name.equals("Pragma", ignoreCase = true) ||
-                name.equals("Expires", ignoreCase = true)
-
-        override fun setHeader(name: String, value: String) { if (!isCacheHeader(name)) super.setHeader(name, value) }
-        override fun addHeader(name: String, value: String) { if (!isCacheHeader(name)) super.addHeader(name, value) }
-        override fun setIntHeader(name: String, value: Int) {
-            if (!isCacheHeader(name)) super.setIntHeader(name, value)
-        }
-
-        override fun addIntHeader(name: String, value: Int) {
-            if (!isCacheHeader(name)) super.addIntHeader(name, value)
-        }
-
-        override fun setDateHeader(name: String, date: Long) {
-            if (!isCacheHeader(name)) super.setDateHeader(name, date)
-        }
-
-        override fun addDateHeader(name: String, date: Long) {
-            if (!isCacheHeader(name)) super.addDateHeader(name, date)
-        }
-
-        private fun applyCacheHeaders() {
-            super.setHeader("Cache-Control", cacheControl)
-        }
-
-        override fun getOutputStream(): ServletOutputStream {
-            applyCacheHeaders()
-            return super.getOutputStream()
-        }
-
-        override fun getWriter(): PrintWriter {
-            applyCacheHeaders()
-            return super.getWriter()
         }
     }
