+++ src/test/kotlin/nl/info/zac/app/filter/cache/StaticCacheFilterTest.kt
@@ -2,75 +2,25 @@
  * SPDX-FileCopyrightText: 2026 INFO.nl
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package nl.info.zac.util
+package nl.info.zac.app.filter.cache
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeSameInstanceAs
+import io.mockk.checkUnnecessaryStub
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.runs
 import io.mockk.slot
 import jakarta.servlet.FilterChain
-import jakarta.servlet.ServletOutputStream
-import jakarta.servlet.WriteListener
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import java.io.PrintWriter
-import java.io.StringWriter
-import java.util.Locale
 
-/**
- * Minimal HttpServletResponse stub that records headers and exposes an output stream,
- * allowing us to verify the headers set by [StaticResourceCacheFilter]'s response wrapper.
- */
-private class StubHttpServletResponse : HttpServletResponse {
-    val headers = mutableMapOf<String, String>()
-    private var committed = false
+class StaticCacheFilterTest : BehaviorSpec({
+    beforeEach { checkUnnecessaryStub() }
 
-    override fun setHeader(name: String, value: String) { headers[name] = value }
-    override fun addHeader(name: String, value: String) { headers[name] = value }
-    override fun containsHeader(name: String) = headers.containsKey(name)
-    override fun getHeader(name: String) = headers[name]
-    override fun getHeaders(name: String) = headers[name]?.let { listOf(it) } ?: emptyList()
-    override fun getHeaderNames() = headers.keys
-    override fun isCommitted() = committed
-    override fun getOutputStream() = object : ServletOutputStream() {
-        override fun isReady() = true
-        override fun setWriteListener(listener: WriteListener?) = Unit
-        override fun write(b: Int) { committed = true }
-    }
-    override fun getWriter() = PrintWriter(StringWriter()).also { committed = true }
-    override fun setStatus(sc: Int) = Unit
-    override fun setContentType(type: String?) = Unit
-    override fun setContentLength(len: Int) = Unit
-    override fun setContentLengthLong(len: Long) = Unit
-    override fun setCharacterEncoding(charset: String?) = Unit
-    override fun setBufferSize(size: Int) = Unit
-    override fun setIntHeader(name: String, value: Int) { headers[name] = value.toString() }
-    override fun addIntHeader(name: String, value: Int) { headers[name] = value.toString() }
-    override fun setDateHeader(name: String, date: Long) { headers[name] = date.toString() }
-    override fun addDateHeader(name: String, date: Long) { headers[name] = date.toString() }
-    override fun getBufferSize() = 0
-    override fun flushBuffer() { committed = true }
-    override fun resetBuffer() = Unit
-    override fun reset() = Unit
-    override fun getCharacterEncoding() = "UTF-8"
-    override fun getContentType() = "text/plain"
-    override fun getLocale(): Locale = Locale.getDefault()
-    override fun setLocale(loc: Locale?) = Unit
-    override fun sendError(sc: Int) = Unit
-    override fun sendError(sc: Int, msg: String?) = Unit
-    override fun sendRedirect(location: String?) = Unit
-    override fun encodeURL(url: String?) = url
-    override fun encodeRedirectURL(url: String?) = url
-    override fun getStatus() = 200
-    override fun addCookie(cookie: jakarta.servlet.http.Cookie?) = Unit
-}
-
-class StaticResourceCacheFilterTest : BehaviorSpec({
-    val filter = StaticResourceCacheFilter()
+    val filter = StaticCacheFilter()
 
     fun request(path: String, queryParam: String? = null) = mockk<HttpServletRequest> {
         every { requestURI } returns path
@@ -84,15 +34,7 @@ class StaticResourceCacheFilterTest : BehaviorSpec({
         }
     }
 
-    fun filterChainThatPassesThrough() = mockk<FilterChain> {
-        every { doFilter(any(), any()) } just runs
-    }
-
-    fun filterChainThatOpensWriter() = mockk<FilterChain> {
-        every { doFilter(any(), any()) } answers { (secondArg() as HttpServletResponse).writer }
-    }
-
-    Given("a hashed JS chunk") {
+Given("a hashed JS chunk") {
         val response = StubHttpServletResponse()
         When("the filter processes the request") {
             filter.doFilter(request("/chunk-A1B2C3D4.js"), response, filterChainThatOpensOutputStream())
@@ -159,7 +101,7 @@ class StaticResourceCacheFilterTest : BehaviorSpec({
     Given("a hashed JS chunk served without opening a stream (e.g. 304/HEAD)") {
         val response = StubHttpServletResponse()
         When("the filter processes the request without accessing the output stream") {
-            filter.doFilter(request("/chunk-A1B2C3D4.js"), response, filterChainThatPassesThrough())
+            filter.doFilter(request("/chunk-A1B2C3D4.js"), response, mockk { every { doFilter(any(), any()) } just runs })
             // Headers are only injected when getOutputStream/getWriter is called. For 304/HEAD
             // responses this is acceptable — the browser already holds the correct immutable
             // Cache-Control from the original 200 response and uses it for subsequent requests.
@@ -172,7 +114,7 @@ class StaticResourceCacheFilterTest : BehaviorSpec({
     Given("a resource served via getWriter") {
         val response = StubHttpServletResponse()
         When("the filter processes the request") {
-            filter.doFilter(request("/index.html"), response, filterChainThatOpensWriter())
+            filter.doFilter(request("/index.html"), response, mockk { every { doFilter(any(), any()) } answers { (secondArg() as HttpServletResponse).writer } })
             Then("Cache-Control is set to no-cache") {
                 response.headers["Cache-Control"] shouldBe "no-cache"
             }
